requirements.txt
autopep8==2.0.2
azure-common==1.1.28
azure-core==1.26.3
azure-cosmos==4.3.1
azure-functions==1.13.3
azure-servicebus==7.8.3
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pycodestyle==2.10.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.5.0
uamqp==1.6.4
urllib3==1.26.15
